TodoApp/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
decorator==5.1.1
easydict==1.13
ecdsa==0.19.0
executing==2.1.0
fastapi==0.112.2
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
ipython==8.27.0
jedi==0.19.1
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
numpy==2.1.1
packaging==24.1
parso==0.8.4
passlib==1.7.4
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.3
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
stack-data==0.6.3
starlette==0.38.4
traitlets==5.14.3
typing_extensions==4.12.2
uvicorn==0.30.6
wcwidth==0.2.13
python-dotenv==1.0.1